FS#10886 - Hebrew glyphs for 08-Rockfont.bdf
This patch adds Hebrew glyphs for the 08-Rockfont font.
This font is the default font for all targets with LCD_HEIGHT<=64
The glyphs are copied from the Liquid Font Family for Very Small Displays
(Liquid Mean).
Henrik Theiling, the author of this font, explicitly approved to integrate his
font into Rockbox (thanks!).


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24137 a1c6a512-1295-4272-9138-f99709370657