Try to detect Wine.
Make the Windoes binary try to detect if it's running under Wine. Display a
warning if so, as Rockbox Utility is likely to fail in all kind of strange ways
when running on Wine. We do provide Linux binaries, so it's not even a good
idea to try running under Wine at all.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@25471 a1c6a512-1295-4272-9138-f99709370657